Update trades.asm
Starting to balance trades. I liked the idea that Mew should be challenging to get, and having to catch a Dratini, and evolve it twice, is my first attempt to see if this is a fair trade. With that said, Raichu can't evolve, and while I can change the Evolve text the trainer uses, I can just as well change the Pokemon to one that does evolve, since there are only four. Or, well, was, before I changed that too.