Update CeladonCity.asm
I wasn't a fan of the slightly inappropriate remark of the old man near the gym, so I tweaked the dialogue slightly to be a little better. Not sure if I will stick with calling them "gurus" over something like "pros" or "aces", but we only have five letters to use, and there really isn't a lot of good options to use here.