Update FuchsiaCity.asm
The sign trigger is in a mountain. Anyone else think that's strange? I moved the sign to the entrance to the right of the city. I know the city has two entrances, and you could use either one, but we only get one sign, so on the right it goes.