Update CeladonGym.asm
For gyms, I thought it might be a good idea to require the player to battle all the trainers in a gym before battling the gym leader themselves. Most of the other gyms can have simple room changes, but this is the only one I can think of that requires the gym leader and a trainer to switch places, since you don't have a hedge you can cut on the left side of the hedges.